+++ fastapi_opa/opa/opa_middleware.py
@@ -4,6 +4,7 @@ import logging
 from json.decoder import JSONDecodeError
 from typing import List
 from typing import Optional
+from unittest.mock import patch
 
 import requests
 from fastapi.responses import JSONResponse
@@ -38,6 +39,7 @@ class OPAMiddleware:
     async def __call__(
         self, scope: Scope, receive: Receive, send: Send
     ) -> None:
+
         request = Request(scope, receive, send)
 
         if request.method == "OPTIONS":
@@ -95,7 +97,11 @@ class OPAMiddleware:
         if not is_authorized:
             return await self.get_unauthorized_response(scope, receive, send)
 
-        return await self.app(scope, receive, send)
+        # Small hack to avoid reading twice the request's body in the
+        # middleware stack
+        # See https://github.com/tiangolo/fastapi/issues/394 for more details
+        with patch.object(Request, "body", request.body):
+            return await self.app(scope, receive, send)
 
     @staticmethod
     async def get_unauthorized_response(