+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [1.3.2] - 2022-03-10
-- Add the option to skip an injectable for some given endpoints.
+- Add the option to skip some given endpoints (middleware + injectable).
 
 ## [1.3.1] - 2021-09-19