+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [1.3.2] - 2022-03-10
 - Add the option to skip some given endpoints (middleware + injectable).
+- Allow authentication through bearer token
+- Fix a bug with graphql injectable
 
 ## [1.3.1] - 2021-09-19